requirements.txt
python-telegram-bot~=20.0.0
python-dotenv~=0.21.0
emoji~=2.2.0